+++ src/Language/Fortran/Extras/Util.hs
@@ -1,0 +1,8 @@
+module Language.Fortran.Extras.Util where
+
+import qualified Data.Text as Text
+import Data.Text ( Text )
+
+tshow :: Show a => a -> Text
+tshow = Text.pack . show
+