+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+  * Update to fortran-src X.Y.Z
+  * Add helpers for using Fortran 77 include parser with IO actions
+    `Language.Fortran.Extras.withToolOptionsAndProgramOrBlock`
+  * Add `ToJSON` instances for data types in `Language.Fortran.AST`
+
 ## 0.3.0 (2022-02-15)
   * Update to fortran-src 0.9.0
   * Remove `Language.Fortran.Extras.ModFiles`. The functions are available